+++ packages/http/tools/e2e-tests/src/warrior/controllers/WarriorsPatchQueryNamedController.ts
@@ -1,0 +1,19 @@
+import { controller, PATCH, query } from '@inversifyjs/http-core';
+
+import { WarriorWithQuery } from '../models/WarriorWithQuery';
+
+@controller('/warriors')
+export class WarriorsPatchQueryNamedController {
+  @PATCH()
+  public async patchWarrior(
+    @query('filter') filter: string,
+  ): Promise<WarriorWithQuery> {
+    return {
+      damage: 10,
+      filter,
+      health: 100,
+      range: 1,
+      speed: 10,
+    };
+  }
+}