+++ packages/http/tools/e2e-tests/src/warrior/controllers/WarriorsPostParamsController.ts
@@ -1,0 +1,19 @@
+import { controller, params, POST } from '@inversifyjs/http-core';
+
+import { WarriorWithId } from '../models/WarriorWithId';
+
+@controller('/warriors')
+export class WarriorsPostParamsController {
+  @POST('/:id')
+  public async createWarrior(
+    @params() params: { id: string },
+  ): Promise<WarriorWithId> {
+    return {
+      damage: 10,
+      health: 100,
+      id: params.id,
+      range: 1,
+      speed: 10,
+    };
+  }
+}